% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{save_plot}
\alias{save_plot}
\title{Save ggplots with options}
\usage{
save_plot(filename, plot, saveplots = NULL, width = 180,
  height = 180, units = "mm", dpi = 300, ...)
}
\arguments{
\item{filename}{A file path/name wrapped in \code{here::here()} before being
passed to \code{ggsave()}}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{saveplots}{A logical indicating whether to save the plot. Can be
overridden by a global variable named \code{global_saveplots}.}

\item{width}{Defaults to 180 mm}

\item{height}{Defaults to 180 mm}

\item{units}{Defaults to "mm"}

\item{dpi}{Defaults to 300}

\item{...}{Other arguments passed on to the graphics device function,
as specified by \code{device}.}
}
\description{
Save ggplots with options
}
