% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{ggplot2_themes}
\alias{ggplot2_themes}
\alias{theme_paper}
\alias{theme_poster}
\title{Custom ggplot2 themes.}
\usage{
theme_paper(base_size = 10, legend.key.size = unit(0.09, "in"),
  legend.title = element_blank(), axis.title.x = element_blank(),
  axis.text.x = element_text(size = 8), axis.ticks.x = element_blank(),
  axis.line.x = element_line(color = "black"),
  axis.line.y = element_line(color = "black"),
  plot.title = element_text(hjust = 0.5),
  strip.background = element_blank(), ...)

theme_poster(base_size = 15, strip.background = element_blank(),
  panel.background = element_blank(), panel.grid.minor = element_line(color
  = "grey93"), plot.background = element_rect(fill = "transparent", color =
  NA), ...)
}
\description{
Base ggplot2 themes for paper figures and poster figures. All arguments are
passed on to \code{ggplot2::\link[ggplot2]{theme}()}.
}
